2.1 - a quality update
Added new quality options for download video and audio, this ranges from 144p to 4320p (8k) for video, and 64k-320k for audio. Best is still an option.
With 8k being an option, i moved it to the last 2 options because this is a crazy resolution to download in, so... don't complain to me if it eats your data usage.
Also, some videos use 21:9, 32:9 ratios, so keep that in mind when selecting a quality.
I plan on implementing a custom way of doing this later, but I don't see a reason at the moment (other than audio, for people who like to listen to 32k audio)

Re-enabled the tray icon... i forgot about it.

Also in this release:
Changed the args and output folder appending to the arg string.
Added a boolean for use hls for FFmpeg on reddit links.