some changes
require ffmpeg outside of fix vreddit, you'll be warned if it doesn't exist but any previous things that work still will.
protocol groundwork. using "ytdl:" as a protocol will redirect the text to the main form's textbox, and that's it. I plan on allowing multiple instances of youtube-dl-gui for many download queues.